engine/requirements.txt
pillow
numpy
opencv-python
mss
pynput
pyautogui
screeninfo